embeddings/pyproject.toml
[project]
name = "embeddings"
version = "0.1.0"
description = ""
authors = [
    {name = "Rohit",email = "rohit.khazanchi@teck.com"}
]
readme = "README.md"
[tool.poetry.dependencies]
python = "~3.13"
langchain-postgres = "^0.0.13"
langchain = "^0.3.20"
langchain-ollama = "^0.2.3"
pgvector = "^0.3.6"
langchain-community = "^0.3.19"
psycopg = "^3.2.5"
beautifulsoup4 = "^4.13.3"
pandas = "^2.2.3"
langchain-cohere = "^0.4.2"
langgraph = "^0.3.5"
pyspellchecker = "^0.8.2"

[tool.poetry]
packages = [{include = "embeddings", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
